refactor: 모임 상세 관련 컴포넌트 features/group/components로 이동
