+++ src/features/group/components/group-detail-card.tsx
@@ -3,8 +3,8 @@ import { Badge } from '@/components/atoms/badge';
 import { GroupProgress } from '@/components/atoms/group/particiapant-progress';
 import { PositionBadge } from '@/components/molecules/position-badge';
 import { SkillBadge } from '@/components/molecules/skill-badge';
-import { ParticipantListModal } from '@/components/organisms/participant-list-modal';
 import { BookmarkButtonContainer } from '@/features/bookmark/components/bookmark-button-container';
+import { ParticipantListModal } from '@/features/group/components/participant-list-modal';
 import { GroupDetail, GroupTypeName } from '@/types';
 import { Position, Skill } from '@/types/enums';
 import { formatYearMonthDayWithDot } from '@/utils/dateUtils';
